+++ src/Core/Repositories/Query/Exceptions/InvalidRepositoryConstraintException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Zolta\Cqrs\Repositories\Query\Exceptions;
+
+use InvalidArgumentException;
+
+final class InvalidRepositoryConstraintException extends InvalidArgumentException {}